Move PointDeleter contents to Tree
