+++ Rust/src/tree/mod.rs
@@ -9,8 +9,5 @@ pub use cut::Cut;
 mod node;
 pub use node::{Internal, Leaf, Node};
 
-mod point_deleter;
-pub use point_deleter::{DeleteResult, PointDeleter};
-
 mod tree;
-pub use tree::{AddResult, NodeTraverser, Tree};
+pub use tree::{AddResult, DeleteResult, NodeTraverser, Tree};
